fix: show nft avatar
